🍫 (#54) - move style to each component
