+++ nutella/src/components/MarkdownEditor/Popup/index.tsx
@@ -1,5 +1,5 @@
 import { FC, useContext, useMemo } from "react";
-import { MarkdownContext } from "../../context/MarkdownCotext";
+import { MarkdownContext } from "../../../context/MarkdownCotext";
 import * as S from "./styles";
 
 interface PopupItem {
